Testing transformation in runtime.
